refactor: wraper lexical text component for pbtext and petext compositions
